pyproject.toml
[tool.poetry]
name = "bitly-cli"
version = "0.1.0"
description = ""
authors = ["Lev Bazarnov <16255087+lbazarnov@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "bitly_cli"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
prompt = "^0.4.1"

[tool.poetry.scripts]
bitly-cli = "bitly_cli.url_shortener:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
